teams, user integration, improvements, reservation on labs

+++ src/main/java/tr/com/orioninc/laborant/app/service/AdminService.java
@@ -5,12 +5,13 @@ import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import tr.com.orioninc.laborant.app.model.Lab;
+import tr.com.orioninc.laborant.app.model.Team;
 import tr.com.orioninc.laborant.app.repository.LabRepository;
+import tr.com.orioninc.laborant.app.repository.TeamRepository;
 import tr.com.orioninc.laborant.exception.custom.AlreadyExistsException;
-import tr.com.orioninc.laborant.exception.custom.NotConnectedException;
 import tr.com.orioninc.laborant.exception.custom.NotFoundException;
-import tr.com.orioninc.laborant.security.model.User;
-import tr.com.orioninc.laborant.security.repository.UserRepository;
+import tr.com.orioninc.laborant.app.model.User;
+import tr.com.orioninc.laborant.app.repository.UserRepository;
 
 import java.io.IOException;
 import java.net.InetAddress;
@@ -25,33 +26,7 @@ public class AdminService {
 
     private LabRepository labRepo;
     private UserRepository userRepo;
-
-    public String addNewLab(String labName, String username, String password, String host, Integer port) throws IOException {
-        log.debug("[addNewLab] called");
-        Lab searchLab = labRepo.findByLabName(labName);
-        if (labName == null || labName.isEmpty() || username == null || username.isEmpty() || password == null || password.isEmpty() || host == null || host.isEmpty() || port == null) {
-            throw new IllegalArgumentException("Lab credentials cannot be empty");
-        } else {
-            if (Objects.isNull(searchLab)) {
-                Lab searchHostUserPair = labRepo.findByUserNameAndHost(username, host);
-                if (Objects.isNull(searchHostUserPair)) {
-                    Lab labToBeAdded = new Lab(labName, username, password, host, port);
-                    if (isLabReachable(labToBeAdded.getHost(), 500)) {
-                        labToBeAdded = labRepo.save(labToBeAdded);
-                        log.info("[addNewLab] added new lab named: {}", labToBeAdded.getLabName());
-                        return "Successfully added the new lab named " + labToBeAdded.getLabName() + " to the database";
-                    }
-                } else {
-                    log.warn("[addNewLab] Lab with username {} and host {} already exists in database", username, host);
-                    return "There is already a pair in the database with username: " + username + " and host: " + host;
-                }
-            } else {
-                log.warn("[addNewLab] Lab with name {} already exists in database", labName);
-                return "There is already a lab named " + labName + " in the database. Try again";
-            }
-            return "Lab could not be added to the database";
-        }
-    }
+    private TeamRepository teamRepo;
 
     public List<Lab> getAllLabs() {
         if (labRepo.findAll().isEmpty()) {
@@ -63,10 +38,80 @@ public class AdminService {
         }
     }
 
-    public String assignUserToLab(String userName, String labName) {
-log.debug("[assignUserToLab] called");
+    public String assignLabToTeam(String labName, String teamName) {
+        log.debug("[assignLabToTeam] called");
+        Lab lab = labRepo.findByLabName(labName);
+        Team team = teamRepo.findByName(teamName);
+        if(Objects.equals(labName,"") || labName == null){
+            log.error("[assignLabToTeam] Lab name is empty");
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
+        if(Objects.equals(teamName,"") || teamName == null){
+            log.error("[assignLabToTeam] Team name is empty");
+            throw new IllegalArgumentException("Team name cannot be empty");
+        }
+        if (Objects.isNull(lab)) {
+            log.error("[assignLabToTeam] Lab not found");
+            throw new NotFoundException("Lab not found");
+        }
+        else if (Objects.isNull(team)) {
+            log.error("[assignLabToTeam] Team not found");
+            throw new NotFoundException("Team not found");
+        }
+        else if (team.getLabs().contains(lab)) {
+            log.error("[assignLabToTeam] Lab {} already assigned to team {}", labName, teamName);
+            throw new AlreadyExistsException("Lab already assigned to team");
+        }
+        else {
+            team.getLabs().add(lab);
+            teamRepo.save(team);
+            log.info("[assignLabToTeam] Lab {} assigned to team {}", labName, teamName);
+            return "Lab assigned to team";
+        }
+    }
+
+    public String unassignLabFromTeam(String labName, String teamName) {
+        log.debug("[unassignTeamFromLab] called");
+        Lab lab = labRepo.findByLabName(labName);
+        Team team = teamRepo.findByName(teamName);
+        if(Objects.equals(labName,"") || labName == null){
+            log.error("[unassignLabFromTeam] Lab name is empty");
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
+        if(Objects.equals(teamName,"") || teamName == null){
+            log.error("[unassignLabFromTeam] Team name is empty");
+            throw new IllegalArgumentException("Team name cannot be empty");
+        }
+        if (Objects.isNull(lab)) {
+            log.error("[unassignLabFromTeam] Lab not found");
+            throw new NotFoundException("Lab not found");
+        }
+        else if (Objects.isNull(team)) {
+            log.error("[unassignLabFromTeam] Team not found");
+            throw new NotFoundException("Team not found");
+        }
+        else if (!team.getLabs().contains(lab)) {
+            log.error("[unassignLabFromTeam] Lab {} not assigned to team {}", labName, teamName);
+            throw new NotFoundException("Lab not assigned to team");
+        }
+        else {
+            team.getLabs().remove(lab);
+            teamRepo.save(team);
+            log.info("[unassignLabFromTeam] Lab unasigned from team");
+            return "Lab unassigned from team";
+        }
+    }
+
+    public String assignLabToUser(String userName, String labName) {
+        log.debug("[assignUserToLab] called");
         User user = userRepo.findByUsername(userName);
         Lab lab = labRepo.findByLabName(labName);
+        if (Objects.equals(userName, "") || userName == null){
+            throw new IllegalArgumentException("Username cannot be empty");
+        }
+        if (Objects.equals(labName, "") || labName == null){
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
         if (Objects.isNull(user)) {
             log.warn("[assignUserToLab] User with username {} not found in database", userName);
             throw new NotFoundException("User with username " + userName + " not found in database");
@@ -76,8 +121,14 @@ log.debug("[assignUserToLab] called");
         } else {
             if (user.getLabs().contains(lab)) {
                 log.warn("[assignUserToLab] User with username {} already assigned to lab with name {}", userName, labName);
-                throw new AlreadyExistsException("User with username " + userName + " already assigned to lab with name " + labName);
+                throw new AlreadyExistsException(userName + " already assigned to lab " + labName);
             } else {
+                for (User u : userRepo.findAll()) {
+                    if (u.getLabs().contains(lab)) {
+                        log.warn("[assignUserToLab] Lab with name {} already assigned to user with username {}", labName, u.getUsername());
+                        throw new AlreadyExistsException("Lab with name " + labName + " already assigned to user with username " + u.getUsername());
+                    }
+                }
                 user.getLabs().add(lab);
                 userRepo.save(user);
                 log.info("[assignUserToLab] User with username {} assigned to lab with name {}", userName, labName);
@@ -86,10 +137,16 @@ log.debug("[assignUserToLab] called");
         }
     }
 
-    public String unassignUserFromLab(String userName, String labName) {
+    public String unassignLabFromUser(String userName, String labName) {
         log.debug("[unassignUserFromLab] called");
         User user = userRepo.findByUsername(userName);
         Lab lab = labRepo.findByLabName(labName);
+        if (Objects.equals(userName, "") || userName == null){
+            throw new IllegalArgumentException("Username cannot be empty");
+        }
+        if (Objects.equals(labName, "") || labName == null){
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
         if (Objects.isNull(user)) {
             log.warn("[unassignUserFromLab] User with username {} not found in database", userName);
             throw new NotFoundException("User with username " + userName + " not found in database");
@@ -127,29 +184,14 @@ log.debug("[assignUserToLab] called");
         }
     }
 
-
-
-
-
-    public String deleteLab(String labName) {
-        log.debug("[deleteLab] called");
-        Lab labToBeDeleted = labRepo.findByLabName(labName);
-        if (Objects.isNull(labToBeDeleted)) {
-            log.warn("[deleteLab] no lab in the database named: {}", labName);
-            return "There isn't a lab named " + labName + " found in the database to be deleted";
-        } else {
-            labRepo.delete(labToBeDeleted);
-            log.info("[deleteLab] deleting new lab named: {}", labToBeDeleted.getLabName());
-            return "Lab named " + labName + " is successfully deleted from the database";
-        }
-    }
-
     public Lab findLabByName(String labName) {
+        if (Objects.equals(labName, "") || labName == null){
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
         log.info("[findLabByName] called");
         return labRepo.findByLabName(labName);
     }
 
-    // REST IMPLEMENTATION
     public Lab updateLabByName(String labName, Lab lab) {
         log.debug("[updateLabByName] called");
         Lab labToBeUpdated = findLabByName(labName);
@@ -182,14 +224,9 @@ log.debug("[assignUserToLab] called");
             if (Objects.isNull(searchLab)) {
                 Lab searchHostUserPair = labRepo.findByUserNameAndHost(lab.getUserName(), lab.getHost());
                 if (Objects.isNull(searchHostUserPair)) {
-                    if (isLabReachable(lab.getHost(), 500)) {
                         lab = labRepo.save(lab);
                         log.info("[addNewLab] adding new lab named: {}", lab.getLabName());
                         return lab;
-                    } else {
-                        log.error("[addNewLab] Cannot add lab. Lab is not reachable");
-                        throw new NotConnectedException("Cannot add lab. The lab is not reachable");
-                    }
                 } else {
                     log.warn("[addNewLab] there is already a pair in the database with username: {} and host: {}", lab.getUserName(), lab.getHost());
                     throw new AlreadyExistsException("There is already a pair in the database with username: " + lab.getUserName() + " and host: " + lab.getHost());
@@ -202,6 +239,9 @@ log.debug("[assignUserToLab] called");
     }
 
     public Lab getLab(String labName) {
+        if (Objects.equals(labName, "") || labName == null){
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
         if (findLabByName(labName) == null) {
             log.warn("[getLab] No lab found in database named {}", labName);
             throw new NotFoundException("There is no lab in database named " + labName);
@@ -212,14 +252,26 @@ log.debug("[assignUserToLab] called");
     }
 
     public Lab deleteLabByName(String labName) {
-        log.debug("[deleteLab] called");
+        log.debug("[deleteLabByName] called");
+        if (Objects.equals(labName, "") || labName == null){
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
         Lab labToBeDeleted = labRepo.findByLabName(labName);
         if (Objects.isNull(labToBeDeleted)) {
-            log.warn("[deleteLab] no lab in the database named: {}", labName);
+            log.warn("[deleteLabByName] no lab in the database named: {}", labName);
             throw new NotFoundException("There isn't a lab named " + labName + " found in the database to be deleted");
         } else {
+            if (!userRepo.findAll().isEmpty()) {
+                for (User user : userRepo.findAll()) {
+                    if (user.getLabs().contains(labToBeDeleted)) {
+                        user.getLabs().remove(labToBeDeleted);
+                        userRepo.save(user);
+                        log.info("[deleteLabByName] User with username {} unassigned from lab with name {}", user.getUsername(), labName);
+                    }
+                }
+            }
             labRepo.delete(labToBeDeleted);
-            log.info("[deleteLab] deleting new lab named: {}", labToBeDeleted.getLabName());
+            log.info("[deleteLabByName] deleting new lab named: {}", labToBeDeleted.getLabName());
             return labToBeDeleted;
         }
     }
@@ -240,6 +292,64 @@ log.debug("[assignUserToLab] called");
             }
         }
 
+
+    public Lab reserveLab(String labName) {
+        log.debug("[reserveLab] called");
+        if (Objects.equals(labName, "") || labName == null){
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
+        Lab labToBeReserved = labRepo.findByLabName(labName);
+        if (Objects.isNull(labToBeReserved)) {
+            log.warn("[reserveLab] no lab in the database named: {}", labName);
+            throw new NotFoundException("There isn't a lab named " + labName + " found in the database to be reserved");
+        } else {
+            if (labToBeReserved.getReserved() == null){
+                labToBeReserved.setReserved(true);
+                labRepo.save(labToBeReserved);
+                log.info("[reserveLab] lab with name {} is reserved", labName);
+                return labToBeReserved;
+            }
+            else {
+                if (labToBeReserved.getReserved()) {
+                    log.warn("[reserveLab] lab with name {} is already reserved", labName);
+                    throw new AlreadyExistsException("Lab with name " + labName + " is already reserved");
+                } else {
+                    labToBeReserved.setReserved(true);
+                    labRepo.save(labToBeReserved);
+                    log.info("[reserveLab] lab with name {} is reserved", labName);
+                    return labToBeReserved;
+                }
+            }
+        }
+    }
+    public Lab unreserveLab (String labName){
+        log.debug("[unreserveLab] called");
+        if (Objects.equals(labName, "") || labName == null){
+            throw new IllegalArgumentException("Lab name cannot be empty");
+        }
+        Lab labToBeUnreserved = labRepo.findByLabName(labName);
+        if (Objects.isNull(labToBeUnreserved)) {
+            log.warn("[unreserveLab] no lab in the database named: {}", labName);
+            throw new NotFoundException("There isn't a lab named " + labName + " found in the database to be unreserved");
+        } else {
+            if (labToBeUnreserved.getReserved() == null){
+                labToBeUnreserved.setReserved(false);
+                labRepo.save(labToBeUnreserved);
+                throw new AlreadyExistsException("Lab with name " + labName + " is not reserved already");
+            }
+            else {
+                if (labToBeUnreserved.getReserved()) {
+                    labToBeUnreserved.setReserved(false);
+                    labRepo.save(labToBeUnreserved);
+                    log.info("[unreserveLab] lab with name {} is unreserved", labName);
+                    return labToBeUnreserved;
+                } else {
+                    log.warn("[unreserveLab] lab with name {} is already unreserved", labName);
+                    throw new AlreadyExistsException("Lab with name " + labName + " is already unreserved");
+                }
+            }
+        }
+    }
 
 
 }
